irqchip/bcm2836: Move SMP startup code to arch/arm
One of the RPi-2/3 irqchip's key features is that it contains some
SMP startup code for the 32bit ARM architecture version. The only
reason I can imagine for this is "RPi is special".

Let's move this code where it belongs (in the platform support code),
creating a shared include file for this purpose.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Fixes: 41f4988cc287 ("irqchip/bcm2836: Add SMP support for the 2836")
Reviewed-by: Phil Elwell <phil@raspberrypi.org>
Tested-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Stefan Wahren <stefan.wahren@i2se.com>